Update dependencies
Mostly needed for typescript-to-lua version update so that it can include lua files in the bundled output. The current version we use doesn't do it.

Some changes to wc3 types based on the latest version of the https://github.com/cipherxof/wc3-ts-template

Update to module resolution is due to dependencies requiring node resolution, which requires us changing some imports to use node resolution.