Add new W3CData and W3CChecksum lua libraries
W3CData is used for the compression and decompression of data, as well
as the registering of schemas for event types that data is used for.

W3CChecksum is a CRC32 implementation that can be used to continuously
update a CRC value. This can then be passed to W3CData to create
checksum packets