Add creep loot indicator and creep loot preview (#44)
* Import indicator

* Add new features to `-commands` msg

* Add README short summary

* clarification

* Remove `./dist` at `compileMap()`

Sometimes non clean dist causes issues when modifying a map in WE and rerunning `compileMap()` (e.g. using `test` build)

* Fix indicator showing on invisible units

* Fix indicator too low on Murloc Flesheater

* Fix "Elder Jungle Stalker" missing indicator

"Elder Jungle Stalker" on 1v1_WarHail_v1.1.w3x.
The data for position is not always precise, so we search in a bigger radius and pick the closest unit

* comment `print`

* Fix updateMaps.sh

* Fix loot indicator appearing on peasants if starting location is on a creep camp

* Fix a crash caused by spawning a unit at world bounds

On WitheringFields, when creating a unit at WorldBounds (minX, minY) causes a crash.

Not sure why only on this map, but hopefully it is related to being part of non-playable part of the map. Changed spawning at (0,0) instead.

* Fix temp unit creation is causing minimap creep camp icon to appear

* Add `Thanks` section

* Remove cleaning dist because it is modified at BUILD step (copy resources before compile map).

...

* Add `Known issues` section

* Clarify loot preview command

* Remove "print loot" on ctrl+click (if pressing ctrl+enter, ctrl will 'stuck' - will be considered pressed)

* change commands names to "-loot*"

* Reformat code

* Add more known issues

* Fix `indicator height` on Infernal

* Add `observer mode` section to readme

* Change to Kenshin icons

* cleanup

* Use fixed war3-transformer from NPM